+++ crates/alkahest-renderer/src/ecs/utility.rs
@@ -1,31 +1,50 @@
 use std::f32::consts::PI;
 
 use bevy_ecs::{
+    bundle::Bundle,
     entity::Entity,
     prelude::Component,
     system::{In, Query, Res, ResMut},
 };
 use destiny_pkg::TagHash;
+use ecolor::Rgba;
 use glam::Vec3;
 
 use super::{
-    common::{Icon, Label},
+    common::{Icon, Label, Mutable, RenderCommonBundle},
+    route::{Route, RouteNode},
+    tags::{NodeFilter, Tags},
     visibility::VisibilityHelper,
     MapInfo,
 };
 use crate::{
-    ecs::{resources::SelectedEntity, transform::Transform, visibility::ViewVisibility},
-    icons::{ICON_MAP_MARKER_PATH, ICON_RULER_SQUARE, ICON_SIGN_POLE, ICON_SPHERE},
+    ecs::{
+        hierarchy::Children, resources::SelectedEntity, transform::Transform,
+        visibility::ViewVisibility,
+    },
+    icons::{ICON_RULER_SQUARE, ICON_SIGN_POLE, ICON_SPHERE},
     renderer::{LabelAlign, Renderer, RendererShared},
     util::{
         color::{Color, ColorExt, Hsv},
         text::prettify_distance,
     },
 };
-
 pub trait Utility {
     fn icon() -> Icon;
+    fn label(str: &str) -> Label {
+        Label::from(str)
+    }
     fn default_label() -> Label;
+}
+
+#[derive(Bundle)]
+pub struct UtilityCommonBundle {
+    pub label: Label,
+    pub icon: Icon,
+    pub filter: NodeFilter,
+    pub tags: Tags,
+    pub mutable: Mutable,
+    pub render_common: RenderCommonBundle,
 }
 
 #[derive(Component)]
@@ -92,7 +111,7 @@ impl Default for Sphere {
 
 impl Utility for Sphere {
     fn default_label() -> Label {
-        Label::new_default("Sphere")
+        Label::new_default("Sphere").with_offset(0.0, 0.0, -1.0)
     }
 
     fn icon() -> Icon {
@@ -121,7 +140,7 @@ impl Default for Beacon {
 
 impl Utility for Beacon {
     fn default_label() -> Label {
-        Label::new_default("Beacon")
+        Label::new_default("Beacon").with_offset(0.0, 0.0, -0.5)
     }
 
     fn icon() -> Icon {
@@ -129,61 +148,7 @@ impl Utility for Beacon {
     }
 }
 
-pub struct RouteNode {
-    pub pos: Vec3,
-    pub map_hash: Option<TagHash>,
-    pub is_teleport: bool,
-    pub label: Option<String>,
-}
-
-impl Default for RouteNode {
-    fn default() -> Self {
-        Self {
-            pos: Vec3::ZERO,
-            map_hash: None,
-            is_teleport: false,
-            label: None,
-        }
-    }
-}
-
-#[derive(Component)]
-pub struct Route {
-    pub path: Vec<RouteNode>,
-    pub color: Color,
-    pub rainbow: bool,
-    pub speed_multiplier: f32,
-    pub scale: f32,
-    pub marker_interval: f32,
-    pub show_all: bool,
-    pub activity_hash: Option<TagHash>,
-}
-
-impl Default for Route {
-    fn default() -> Self {
-        Self {
-            path: vec![],
-            color: Color::WHITE,
-            rainbow: false,
-            speed_multiplier: 1.0,
-            scale: 1.0,
-            marker_interval: 0.0,
-            show_all: false,
-            activity_hash: None,
-        }
-    }
-}
-
-impl Utility for Route {
-    fn icon() -> Icon {
-        Icon::Unicode(ICON_MAP_MARKER_PATH)
-    }
-
-    fn default_label() -> Label {
-        Label::new_default("Route")
-    }
-}
-
+#[allow(clippy::too_many_arguments)]
 pub fn draw_utilities_system(
     In(renderer): In<RendererShared>,
     map_info: Option<Res<MapInfo>>,
@@ -191,48 +156,51 @@ pub fn draw_utilities_system(
     q_ruler: Query<(Entity, &Ruler, Option<&ViewVisibility>)>,
     q_sphere: Query<(Entity, &Transform, &Sphere, Option<&ViewVisibility>)>,
     q_beacon: Query<(Entity, &Transform, &Beacon, Option<&ViewVisibility>)>,
-    q_route: Query<(Entity, &Route, Option<&ViewVisibility>)>,
+    q_route: Query<(Entity, &Route, &Children, Option<&ViewVisibility>)>,
+    q_route_node: Query<(Entity, &Transform, &RouteNode)>,
 ) {
     for (e, ruler, vis) in q_ruler.iter() {
         if vis.is_visible(renderer.active_view) {
-            draw_ruler(&renderer, ruler, Some(e), &selected);
+            draw_ruler(&renderer, ruler, e, &selected);
         }
     }
 
     for (e, transform, sphere, vis) in q_sphere.iter() {
         if vis.is_visible(renderer.active_view) {
-            draw_sphere(&renderer, transform, sphere, Some(e), &selected);
+            draw_sphere(&renderer, transform, sphere, e, &selected);
         }
     }
 
     for (e, transform, beacon, vis) in q_beacon.iter() {
         if vis.is_visible(renderer.active_view) {
-            draw_beacon(&renderer, transform, beacon, Some(e), &selected);
+            draw_beacon(&renderer, transform, beacon, e, &selected);
         }
     }
-
-    for (e, route, vis) in q_route.iter() {
+    for (e, route, children, vis) in q_route.iter() {
         if vis.is_visible(renderer.active_view) {
             if let Some(map_info) = &map_info {
-                draw_route(&renderer, route, Some(e), map_info.map_hash, &selected);
+                draw_route(
+                    &renderer,
+                    route,
+                    children,
+                    &q_route_node,
+                    e,
+                    map_info.map_hash,
+                    &selected,
+                );
             }
         }
     }
 }
 
-fn draw_ruler(
-    renderer: &Renderer,
-    ruler: &Ruler,
-    entity: Option<Entity>,
-    selected: &SelectedEntity,
-) {
+fn draw_ruler(renderer: &Renderer, ruler: &Ruler, entity: Entity, selected: &SelectedEntity) {
     let color = if ruler.rainbow {
         Color::from(*Hsv::rainbow())
     } else {
         ruler.color
     };
 
-    let color = selected.select_fade_color(color, entity);
+    let color = selected.select_fade_color(color, Some(entity));
 
     renderer.immediate.cross(ruler.start, ruler.scale, color);
     renderer.immediate.cross(ruler.end, ruler.scale, color);
@@ -356,7 +324,7 @@ fn draw_sphere(
     renderer: &Renderer,
     transform: &Transform,
     sphere: &Sphere,
-    entity: Option<Entity>,
+    entity: Entity,
     selected: &SelectedEntity,
 ) {
     let color = if sphere.rainbow {
@@ -365,7 +333,7 @@ fn draw_sphere(
         sphere.color
     };
 
-    let color = selected.select_fade_color(color, entity);
+    let color = selected.select_fade_color(color, Some(entity));
 
     let color_opaque = color.to_opaque();
     let cross_color = color_opaque.invert().keep_bright();
@@ -398,7 +366,7 @@ fn draw_beacon(
     renderer: &Renderer,
     transform: &Transform,
     beacon: &Beacon,
-    entity: Option<Entity>,
+    entity: Entity,
     selected: &SelectedEntity,
 ) {
     const BEAM_HEIGHT: f32 = 5000.0;
@@ -411,7 +379,7 @@ fn draw_beacon(
         (150.0 + (renderer.time.load().elapsed() * 2.0 * PI * beacon.freq).sin() * 50.0) / 255.0,
     );
 
-    let color = selected.select_fade_color(color, entity);
+    let color = selected.select_fade_color(color, Some(entity));
 
     renderer.immediate.sphere(
         transform.translation,
@@ -443,63 +411,53 @@ fn draw_beacon(
 fn draw_route(
     renderer: &Renderer,
     route: &Route,
-    entity: Option<Entity>,
+    children: &Children,
+    q_route_node: &Query<(Entity, &Transform, &RouteNode)>,
+    entity: Entity,
     current_hash: TagHash,
     selected: &SelectedEntity,
 ) {
     let color = if route.rainbow {
-        selected.select_fade_color(Color::from(*Hsv::rainbow()), entity)
+        selected.select_fade_color(Color::from(*Hsv::rainbow()), Some(entity))
     } else {
-        selected.select_fade_color(route.color, entity)
+        selected.select_fade_color(route.color, Some(entity))
     };
 
-    const BASE_RADIUS: f32 = 0.1;
     let mut prev_is_local = false;
-    for i in 0..route.path.len() {
-        if let Some(node) = route.path.get(i) {
+    for i in 0..children.0.len() {
+        if let Some(node_e) = children.0.get(i) {
+            let Ok((_, pos, node)) = q_route_node.get(*node_e) else {
+                return;
+            };
+            let next_node = children
+                .0
+                .get(i + 1)
+                .and_then(|e| match q_route_node.get(*e) {
+                    Ok((_, t, n)) => Some((t, n)),
+                    Err(_) => None,
+                });
+
             let node_is_local = node.map_hash.map_or(true, |h| h == current_hash);
-            let next_node = route.path.get(i + 1);
-            let next_is_local = next_node.map_or(false, |node| {
-                node.map_hash.map_or(false, |h| h == current_hash)
+            let next_is_local = next_node.map_or(false, |(_, n)| {
+                n.map_hash.map_or(false, |h| h == current_hash)
             });
 
-            if !node_is_local {
-                if prev_is_local || next_is_local || route.show_all {
-                    draw_sphere_skeleton(renderer, node.pos, BASE_RADIUS * route.scale, 2, color);
-                }
-            } else {
-                renderer.immediate.sphere(
-                    node.pos,
-                    BASE_RADIUS * route.scale,
-                    color,
-                    //DebugDrawFlags::DRAW_NORMAL,
-                    //None,
-                );
+            if route.show_all || prev_is_local || node_is_local || next_is_local {
+                draw_route_node(renderer, route, node, pos, color, *node_e, current_hash);
             }
 
-            if node_is_local || prev_is_local || next_is_local {
-                if let Some(label) = node.label.as_ref() {
-                    renderer.immediate.label(
-                        label.to_string(),
-                        node.pos + route.scale / 2.0 * Vec3::Z,
-                        LabelAlign::CENTER_BOTTOM,
-                        Color::WHITE,
-                    );
-                }
-            }
             prev_is_local = node_is_local;
 
-            if next_node.is_some() {
-                let next_node = next_node.unwrap();
-                let segment_length = (next_node.pos - node.pos).length();
+            if let Some((next_pos, next_node)) = next_node {
+                let segment_length = (next_pos.translation - pos.translation).length();
 
                 if !(route.show_all || node_is_local || next_is_local) {
                     continue;
                 }
 
                 renderer.immediate.line_dotted(
-                    next_node.pos,
-                    node.pos,
+                    next_pos.translation,
+                    pos.translation,
                     color,
                     color,
                     1.0,
@@ -519,7 +477,8 @@ fn draw_route(
                     let mut current = 0.0;
                     while current < segment_length {
                         if current > 0.0 {
-                            let pos = node.pos + (next_node.pos - node.pos).normalize() * current;
+                            let pos = pos.translation
+                                + (next_pos.translation - pos.translation).normalize() * current;
 
                             renderer.immediate.sphere(
                                 pos,
@@ -535,9 +494,9 @@ fn draw_route(
                 }
                 //TODO (cohae): Fix this once pick buffer exists
                 // renderer.immediate.cube_extents(
-                //     (node.pos + next_node.pos) / 2.0,
+                //     (pos.translation + next_pos.translation) / 2.0,
                 //     Vec3::new(segment_length / 2.0, route.scale / 2.0, route.scale / 2.0),
-                //     Quat::from_rotation_arc(Vec3::X, (next_node.pos - node.pos).normalize()),
+                //     Quat::from_rotation_arc(Vec3::X, (next_pos.translation - pos.translation).normalize()),
                 //     color,
                 //     true,
                 //     DebugDrawFlags::DRAW_PICK,
@@ -545,7 +504,7 @@ fn draw_route(
                 // )
             } else {
                 // renderer.immediate.cube_extents(
-                //     node.pos,
+                //     pos.translation,
                 //     Vec3::new(route.scale / 2.0, route.scale / 2.0, route.scale / 2.0),
                 //     Quat::IDENTITY,
                 //     color,
@@ -555,5 +514,34 @@ fn draw_route(
                 // )
             }
         }
+    }
+}
+
+fn draw_route_node(
+    renderer: &Renderer,
+    route: &Route,
+    node: &RouteNode,
+    pos: &Transform,
+    color: Rgba,
+    _: Entity,
+    current_hash: TagHash,
+) {
+    const BASE_RADIUS: f32 = 0.1;
+    if node.map_hash.map_or(true, |h| h == current_hash) {
+        renderer.immediate.sphere(
+            pos.translation,
+            BASE_RADIUS * route.scale,
+            color,
+            //DebugDrawFlags::DRAW_NORMAL,
+            //None,
+        );
+    } else {
+        draw_sphere_skeleton(
+            renderer,
+            pos.translation,
+            BASE_RADIUS * route.scale,
+            2,
+            color,
+        );
     }
 }