+++ CHANGELOG.md
@@ -51,15 +51,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 - Added console command `set_camera_from_cb12` to load view matrices from a raw view scope buffer
 - Added commands `lock_time` and `unlock_time` to allow fixing the game time to a specific value (eg. for more deterministic image comparisons)
 - Shadow quality option (replaces shadow checkbox)
+- Added Arrow Hotkeys to navigate selected object heirarchies @Froggy618157725
+  in [#36](https://github.com/cohaereo/alkahest/pull/36)
+- Added hotkeys to add to current route. <kbd>+</kbd> adds at the end of the route, or after current node, <kbd>-</kbd> adds before the selected node
 - Content confidentiality dialog on first startup
 - Log unimplemented TFX bytecode ops in the TFX debugger
-
 ### Changed
 
 - Enable SSAO by default
 - Abstract global/fullscreen pipelines
 - Replaced `hecs` with `bevy-ecs`
 - Use bevy-ecs change detection to update cbuffers
+- Improved Route UI. Can now use translation gizmo for small adjustments
 - Enabled fast math for `glam`
 - Various atmosphere-related improvements